fix(social/calendar): consolidate calendar components, fix today pill + tint, ai assist dialog
Issues 3-6 in one PR:

- Issue 3/4: DefaultCell and DnDCell now use p-2 padding (was p-1) and h-6 w-6
  today pill (was h-5 w-5); today cell gets bg-primary/5 tint when not selected.

- Issue 5: Merged MonthCalendar + DayCell into SocialCalendarGrid.tsx.
  CalendarShell now inlines DnDCell (testid calendar-dnd-cell) via renderDay;
  no cell-add-btn anywhere. Deleted MonthCalendar.tsx, DayCell.tsx,
  CalendarCell.tsx, SocialCalendarClient.tsx. E2E tests updated accordingly:
  calendar-cell → calendar-dnd-cell; F-3 test removed.

- Issue 6: AI assistant panel moved from Radix Popover to Dialog
  (max-w-[600px]) so it no longer overflows the composer bounds.
