+++ components/social/composer/ComposerOverlay.tsx
@@ -8,7 +8,7 @@ import { cn } from "@/lib/utils";
 import { ProfileSelector } from "@/components/social/composer/ProfileSelector";
 import { ComposerEditor } from "@/components/social/composer/ComposerEditor";
 import { PreviewCard } from "@/components/social/composer/PreviewCard";
-import { MonthCalendar } from "@/components/social/calendar/MonthCalendar";
+import { SocialCalendarGrid } from "@/components/social/calendar/SocialCalendarGrid";
 import { SchedulingCard, defaultSchedulingCardValue, type SchedulingCardValue } from "@/components/social/composer/SchedulingCard";
 import { UnsavedChangesDialog } from "@/components/social/composer/UnsavedChangesDialog";
 import { ComposerErrorBoundary } from "@/components/social/composer/ComposerErrorBoundary";
@@ -687,7 +687,7 @@ export function ComposerOverlay({
                   />
                 )
               ) : (
-                <MonthCalendar
+                <SocialCalendarGrid
                   companyId={companyId}
                   selectedDate={prefilledDate}
                   highlightPostId={initialDraft?.id}
